Introducing /snap/patches directory
This patch introduces the /snap/patches directory for packagers to place
patches used in the packaging process.  A `patches` part is included to
install these files under $SNAPCRAFT_STAGE/packages to be used by other
parts.

A Sed script is also included to patch the desktop entries' name and
icon.

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>